requirements-dev.txt
# Requirements for Fairlearn development

# Changes to this file (especially in the 'test' section) should be cross-checked against the
# release workflow in
# .github/workflows/release-wheel.yml
# Since that file does its own testing setup

# Required for environment
ruff==0.4.5
requirements-parser
# Need to keep black version consistent in
# requirements-dev.txt
# linting.yml
# .pre-commit-config.yaml
black==24.4.2

# Required for test
pytest
pytest-cov
pytest-mock
pytest-mpl
wheel
lightgbm
xlrd
polars

# Required for documentation
pypandoc
sphinx==7.1.2; python_version=="3.8"
sphinx==7.2.6; python_version>"3.8"
sphinx-gallery>=0.18.0
numpydoc
pydata-sphinx-theme>=0.14.4
sphinx-autodoc-typehints
sphinxcontrib-bibtex
sphinx-prompt
sphinx-copybutton
jupyterlite_sphinx
jupyterlite-pyodide-kernel
packaging
seaborn
torch
sphinx-issues
